pyproject.toml
[project]
name = "mcp.gawron.cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.43.0",
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.10.4",
    "flair>=0.15.1",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-docling>=1.0.0",
    "langchain-openai>=0.3.28",
    "langchain-unstructured>=0.1.6",
    "langchain>=0.3.27",
    "pdfminer>=20191125",
    "python-dotenv>=1.1.1",
    "spacy>=3.8.7",
    "typer>=0.16.0",
]
